fix: refreshing of datalake expired schemas causing inifinite loop (#5530)
# Description

**BUG**: In the `getSchema` method of `schemaV2` if the schema is found
to be expired, it calls the `FetchSchema` method of the corresponding
warehouse. In case of datalake, it calls `GetLocalSchema` of the
uploader which calls the `GetLocalSchema` of schema. This created an
inifinte loop as `GetLocalSchema` again calls `getSchema`

**FIX**: Moved the logic of fetching schema from DB from `schema` to
`uploader` and removed it from the `Handle` interface

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.