fix: group jobs by dest id to transformer from router (#5598)
# Description
OAuth v2 assumes all jobs in batch are related to a single destination
ID, and that only happens if isolation mode is destination, but in
hosted we have isolation mode set to none so we are grouping jobs by
destination before going to transformer.

Resolves INT-3394

## Linear Ticket

https://linear.app/rudderstack/issue/INT-3394

## Security

- [ ] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.

---------

Co-authored-by: Aris Tzoumas <atzoumas@rudderstack.com>