fix: kafka message batching applies to Confluent Cloud and Azure Event Hub (#5699)
# Description

Kafka stream manager is used for `Azure Event Hub` and `Confluent Cloud`
destinations. The problem is message batching directive(configured for
Kafka) is applied across all three as well which can be a problem if
batched messages cannot be accepted further down the pipeline. The same
compression directive still applies to all three.

## Linear Ticket

Resolves PIPE-1982
[Remove kafka batching on transformer and use client batching with a new
config
flag](https://linear.app/rudderstack/issue/PIPE-1982/remove-kafka-batching-on-transformer-and-use-client-batching-with-a)

## Security

- [ ] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.